feat: separate publish button component
