+++ app/layout/projects/show/header/toolbar/publish-btn/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
